Update TTN-TTI-ENCODER - V3 - V4 - version 2
This updated version takes the time of the LNS at the time of the time synchronization. You will need to modify the lknd #30 by changing the offset depending of your time zone location compared to the time of the TTN server (i.e. +4 above is the time difference betwen GMT and Dubai, UAE).